refactor: add imports analyzer class to process all imports
