+++ lib/utils.ts
@@ -1,21 +1,7 @@
-import { FileAnalyzerImportsResult } from './fileAnalyzer';
+import { ImportsAnalyzerResult } from './importsAnalyzer';
 
 export class Utils {
   static isRelative(file: string) {
     return file.startsWith('.');
   }
-
-  static isAllImport(parentImport?: FileAnalyzerImportsResult) {
-    return (
-      !parentImport ||
-      (parentImport !== null &&
-        parentImport.globalRenameImport === null &&
-        parentImport.namedImports === null)
-    );
-  }
-
-  static isRenameGlobalImport(parentImport?: FileAnalyzerImportsResult) {
-    return parentImport && parentImport.globalRenameImport !== null;
-  }
-
 }
