+++ test/fileAnalyzer.spec.ts
@@ -2,155 +2,6 @@ import { assert } from 'chai';
 import { FileAnalyzer } from '../lib/fileAnalyzer';
 
 describe('FileAnalyzer', () => {
-  describe('analyzeImport', () => {
-    const fileAnalyzer = new FileAnalyzer('./contracts/LocalImports.sol');
-
-    it('should throw an exception if import is incorrect', () => {
-      assert.throws(
-        () => fileAnalyzer.analyzeImport('import "filename.sol;'),
-        Error,
-        'Unknown import statement',
-      );
-    });
-
-    it('should analyze global imports', () => {
-      const result = fileAnalyzer.analyzeImport('import "filename.sol";');
-      assert.deepEqual(result, {
-        file: 'filename.sol',
-        importStatement: 'import "filename.sol";',
-        globalRenameImport: null,
-        namedImports: null,
-      });
-    });
-
-    it('should analyze global rename', () => {
-      let result = fileAnalyzer.analyzeImport(
-        'import "filename.sol" as myName;',
-      );
-      assert.deepEqual(
-        result,
-        {
-          file: 'filename.sol',
-          importStatement: 'import "filename.sol" as myName;',
-          globalRenameImport: 'myName',
-          namedImports: null,
-        },
-        'solidity as syntax',
-      );
-
-      result = fileAnalyzer.analyzeImport(
-        'import * as myName from "filename.sol";',
-      );
-      assert.deepEqual(
-        result,
-        {
-          file: 'filename.sol',
-          importStatement: 'import * as myName from "filename.sol";',
-          globalRenameImport: 'myName',
-          namedImports: null,
-        },
-        '* as syntax',
-      );
-    });
-
-    it('should analyze named imports', () => {
-      const result = fileAnalyzer.analyzeImport(
-        'import { A, B as Bingo, C as C1, D } from "filename.sol";',
-      );
-      assert.deepEqual(result, {
-        file: 'filename.sol',
-        importStatement: 'import { A, B as Bingo, C as C1, D } from "filename.sol";',
-        globalRenameImport: null,
-        namedImports: [
-          {
-            name: 'A',
-            as: null,
-          },
-          {
-            name: 'B',
-            as: 'Bingo',
-          },
-          {
-            name: 'C',
-            as: 'C1',
-          },
-          {
-            name: 'D',
-            as: null,
-          },
-        ],
-      });
-    });
-  });
-
-  describe('analyzeImports', () => {
-    const fileAnalyzer = new FileAnalyzer('./contracts/LocalImports.sol');
-
-    it('should analyze all imports', () => {
-      const imports = fileAnalyzer.analyzeImports(`
-        import "filename.sol";
-        import "filename1.sol" as myName;
-        import * as myName from "filename2.sol";
-        import { A, B as Bingo, C as C1, D } from "filename3.sol";
-
-        Some contract text that is not required here
-      `);
-
-      assert.deepEqual(imports, [
-        {
-          file: 'filename.sol',
-          importStatement: 'import "filename.sol";',
-          globalRenameImport: null,
-          namedImports: null,
-        },
-        {
-          file: 'filename1.sol',
-          importStatement: 'import "filename1.sol" as myName;',
-          globalRenameImport: 'myName',
-          namedImports: null,
-        },
-        {
-          file: 'filename2.sol',
-          importStatement: 'import * as myName from "filename2.sol";',
-          globalRenameImport: 'myName',
-          namedImports: null,
-        },
-        {
-          file: 'filename3.sol',
-          importStatement: 'import { A, B as Bingo, C as C1, D } from "filename3.sol";',
-          globalRenameImport: null,
-          namedImports: [
-            {
-              name: 'A',
-              as: null,
-            },
-            {
-              name: 'B',
-              as: 'Bingo',
-            },
-            {
-              name: 'C',
-              as: 'C1',
-            },
-            {
-              name: 'D',
-              as: null,
-            },
-          ],
-        },
-      ]);
-    });
-
-    it('should return empty array if there are no imports', () => {
-      const imports = fileAnalyzer.analyzeImports(`
-        Some contracts without imports
-
-        Some contract text that is not required here
-      `);
-
-      assert.deepEqual(imports, []);
-    });
-  });
 
   describe('analyzeExports', () => {
     const fileAnalyzer = new FileAnalyzer('./contracts/LocalImports.sol');
